refactor: move API types to types/api and update related imports
